* small update in hashmap/hashset code

+++ src/ta3d/src/zuzuf/hashmap.h
@@ -328,7 +328,7 @@ namespace zuzuf
 
 		const HFn hash;
 		size_t h = (hash(key) << 1) & _mask;
-		size_t first_suitable_place = -1;
+		size_t first_suitable_place = size_t(-1);
 		do
 		{
 			switch(usemask[h])
@@ -343,14 +343,14 @@ namespace zuzuf
 				h = (h + 1U) & _mask;
 				continue;
 			case Deleted:
-				if (first_suitable_place == -1)
+				if (first_suitable_place == size_t(-1))
 					first_suitable_place = h;
 				h = (h + 1U) & _mask;
 				continue;
 			}
 			break;
 		} while(true);
-		if (first_suitable_place != -1)
+		if (first_suitable_place != size_t(-1))
 			h = first_suitable_place;
 		++_size;
 		++_used;
@@ -417,7 +417,7 @@ namespace zuzuf
 
 		const HFn hash;
 		size_t h = (hash(key) << 1) & _mask;
-		size_t first_suitable_place = -1;
+		size_t first_suitable_place = size_t(-1);
 		do
 		{
 			switch(usemask[h])
@@ -428,14 +428,14 @@ namespace zuzuf
 				h = (h + 1U) & _mask;
 				continue;
 			case Deleted:
-				if (first_suitable_place == -1)
+				if (first_suitable_place == size_t(-1))
 					first_suitable_place = h;
 				h = (h + 1U) & _mask;
 				continue;
 			}
 			break;
 		} while(true);
-		if (first_suitable_place != -1)
+		if (first_suitable_place != size_t(-1))
 			h = first_suitable_place;
 		++_size;
 		++_used;
